del preprocessing, outsource functions to dlframework
